working tests after big refactor
