+++ contracts/sg-eth-airdrop/src/helpers/constants.rs
@@ -3,3 +3,4 @@ pub const NATIVE_DENOM: &str = "ustars";
 pub const CONTRACT_NAME: &str = "crates.io:sg-eth-airdrop";
 pub const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
 pub const INIT_WHITELIST_REPLY_ID: u64 = 1;
+pub const COLLECTION_WHITELIST: &str = "contract1";
