+++ contracts/sg-eth-airdrop/src/helpers/responses.rs
@@ -1,6 +1,6 @@
+use crate::error::ContractError;
 use crate::state::CONFIG;
-use crate::{build_msg::build_update_minter_address_msg, error::ContractError};
-use cosmwasm_std::{Addr, DepsMut, MessageInfo};
+use cosmwasm_std::{DepsMut, MessageInfo};
 use sg_std::Response;
 
 use crate::build_msg::{build_add_eth_eligible_msg, build_remove_eth_eligible_msg};
@@ -27,7 +27,7 @@ pub fn get_add_eligible_eth_response(
 }
 
 pub fn get_remove_eligible_eth_response(
-    deps: DepsMut,
+    deps: &DepsMut,
     eth_address: String,
 ) -> Result<Response, ContractError> {
     let config = CONFIG.load(deps.storage)?;
