+++ contracts/sg-eth-airdrop/src/testing/constants/claim_constants.rs
@@ -1,5 +1,5 @@
 pub const OWNER: &str = "admin0001";
-pub const AIRDROP_CONTRACT: &str = "contract0";
+pub const AIRDROP_CONTRACT: &str = "contract4";
 pub const MINTER_CONTRACT: &str = "contract1";
 pub const STARGAZE_WALLET_01: &str = "0xstargaze_wallet_01";
 pub const STARGAZE_WALLET_02: &str = "0xstargaze_wallet_02";
