+++ contracts/sg-eth-airdrop/src/testing/constants/collection_constants.rs
@@ -1,0 +1,14 @@
+pub const CREATION_FEE: u128 = 5_000_000_000;
+pub const INITIAL_BALANCE: u128 = 2_000_000_000;
+
+pub const MINT_PRICE: u128 = 100_000_000;
+pub const WHITELIST_AMOUNT: u128 = 66_000_000;
+pub const WL_PER_ADDRESS_LIMIT: u32 = 1;
+pub const MAX_TOKEN_LIMIT: u32 = 10000;
+
+pub const MIN_MINT_PRICE: u128 = 50_000_000;
+pub const AIRDROP_MINT_PRICE: u128 = 0;
+pub const MINT_FEE_BPS: u64 = 1_000; // 10%
+pub const AIRDROP_MINT_FEE_BPS: u64 = 10_000; // 100%
+pub const SHUFFLE_FEE: u128 = 500_000_000;
+pub const MAX_PER_ADDRESS_LIMIT: u32 = 50;
