+++ contracts/sg-eth-airdrop/src/testing/tests/test_claim.rs
@@ -5,16 +5,15 @@ use cosmwasm_std::{Addr, Attribute, Coin, Uint128};
 use ethers_core::rand::thread_rng;
 use ethers_signers::{LocalWallet, Signer};
 
-use crate::tests_folder::shared::{
-    execute_contract_with_msg, get_msg_plaintext, get_signature, get_wallet_and_sig,
-    instantiate_contract_get_app,
+use crate::tests_folder::claim_constants::MINTER_CONTRACT;
+use crate::tests_folder::claim_constants::{
+    AIRDROP_CONTRACT, NATIVE_DENOM, OWNER, STARGAZE_WALLET_01, STARGAZE_WALLET_02,
 };
-
-use crate::tests_folder::constants::{
-    AIRDROP_CONTRACT, NATIVE_DENOM, STARGAZE_WALLET_01, STARGAZE_WALLET_02,
+use crate::tests_folder::setup_contracts::{
+    custom_mock_app, execute_contract_with_msg, instantiate_contract, instantiate_contract_get_app,
 };
-
-use super::constants::MINTER_CONTRACT;
+use crate::tests_folder::setup_minter::configure_minter_with_whitelist;
+use crate::tests_folder::setup_signatures::{get_msg_plaintext, get_signature, get_wallet_and_sig};
 
 #[test]
 fn test_instantiate() {
@@ -26,9 +25,17 @@ fn test_instantiate() {
 fn test_valid_eth_sig_claim() {
     let claim_plaintext = &get_msg_plaintext(STARGAZE_WALLET_01.to_string());
     let (_, eth_sig_str, _, eth_addr_str) = get_wallet_and_sig(claim_plaintext.clone());
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app =
-        instantiate_contract_get_app(vec![eth_addr_str.clone()], 10000, 1, minter_address);
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+
+    instantiate_contract(
+        vec![eth_addr_str.clone()],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
 
     let claim_message = ExecuteMsg::ClaimAirdrop {
         eth_address: eth_addr_str,
@@ -47,7 +54,7 @@ fn test_valid_eth_sig_claim() {
     let expected_attributes = [
         Attribute {
             key: "_contract_addr".to_string(),
-            value: "contract0".to_string(),
+            value: AIRDROP_CONTRACT.to_string(),
         },
         Attribute {
             key: "claimed_amount".to_string(),
@@ -75,28 +82,34 @@ fn test_invalid_eth_sig_claim() {
     let (_, _, _, eth_addr_str) = get_wallet_and_sig(claim_plaintext.clone());
     let (_, eth_sig_str_2, _, _) = get_wallet_and_sig(claim_plaintext.clone());
 
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app =
-        instantiate_contract_get_app(vec![eth_addr_str.clone()], 10000, 1, minter_address);
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+    instantiate_contract(
+        vec![eth_addr_str.clone()],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
 
     let claim_message = ExecuteMsg::ClaimAirdrop {
         eth_address: eth_addr_str,
         eth_sig: eth_sig_str_2,
     };
     let stargaze_wallet_01 = Addr::unchecked(STARGAZE_WALLET_01);
-    let airdrop_contract = Addr::unchecked(AIRDROP_CONTRACT);
     let res = execute_contract_with_msg(
         claim_message,
         &mut app,
         stargaze_wallet_01,
-        airdrop_contract,
+        Addr::unchecked(AIRDROP_CONTRACT),
     )
     .unwrap();
 
     let expected_attributes = [
         Attribute {
             key: "_contract_addr".to_string(),
-            value: "contract0".to_string(),
+            value: AIRDROP_CONTRACT.to_string(),
         },
         Attribute {
             key: "claimed_amount".to_string(),
@@ -123,9 +136,17 @@ fn test_can_not_claim_twice() {
     let claim_plaintext = &get_msg_plaintext(STARGAZE_WALLET_01.to_string());
     let (_, eth_sig_str, _, eth_addr_str) = get_wallet_and_sig(claim_plaintext.clone());
 
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app =
-        instantiate_contract_get_app(vec![eth_addr_str.clone()], 10000, 1, minter_address);
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+    instantiate_contract(
+        vec![eth_addr_str.clone()],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
+
     let claim_message = ExecuteMsg::ClaimAirdrop {
         eth_address: eth_addr_str,
         eth_sig: eth_sig_str,
@@ -143,7 +164,7 @@ fn test_can_not_claim_twice() {
     let expected_attributes = [
         Attribute {
             key: "_contract_addr".to_string(),
-            value: "contract0".to_string(),
+            value: AIRDROP_CONTRACT.to_string(),
         },
         Attribute {
             key: "claimed_amount".to_string(),
@@ -173,7 +194,7 @@ fn test_can_not_claim_twice() {
     let expected_attributes = [
         Attribute {
             key: "_contract_addr".to_string(),
-            value: "contract0".to_string(),
+            value: AIRDROP_CONTRACT.to_string(),
         },
         Attribute {
             key: "claimed_amount".to_string(),
@@ -201,9 +222,17 @@ fn test_claim_one_valid_airdrop() {
     let (_, eth_sig_str, _, eth_addr_str) = get_wallet_and_sig(claim_plaintext.clone());
 
     let stargaze_wallet_01 = Addr::unchecked(STARGAZE_WALLET_01);
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app =
-        instantiate_contract_get_app(vec![eth_addr_str.clone()], 10000, 1, minter_address);
+
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+    instantiate_contract(
+        vec![eth_addr_str.clone()],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
 
     let balances = app
         .wrap()
@@ -238,9 +267,16 @@ fn test_claim_twice_receive_funds_once() {
     let (_, eth_sig_str, _, eth_addr_str) = get_wallet_and_sig(claim_plaintext.clone());
 
     let stargaze_wallet_01 = Addr::unchecked(STARGAZE_WALLET_01);
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app =
-        instantiate_contract_get_app(vec![eth_addr_str.clone()], 10000, 1, minter_address);
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+    instantiate_contract(
+        vec![eth_addr_str.clone()],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
     let balances = app
         .wrap()
         .query_all_balances(stargaze_wallet_01.clone())
@@ -256,7 +292,7 @@ fn test_claim_twice_receive_funds_once() {
         claim_message.clone(),
         &mut app,
         stargaze_wallet_01.clone(),
-        airdrop_contract,
+        airdrop_contract.clone(),
     )
     .unwrap();
 
@@ -269,7 +305,6 @@ fn test_claim_twice_receive_funds_once() {
         amount: Uint128::new(3000),
     }];
     assert_eq!(balances, expected_balance);
-    let airdrop_contract = Addr::unchecked(AIRDROP_CONTRACT);
     let _ = execute_contract_with_msg(
         claim_message,
         &mut app,
@@ -290,8 +325,17 @@ fn test_claim_twice_receive_funds_once() {
 fn test_ineligible_does_not_receive_funds() {
     let claim_plaintext = &get_msg_plaintext(STARGAZE_WALLET_01.to_string());
     let (_, _, _, eth_addr_str) = get_wallet_and_sig(claim_plaintext.clone());
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app = instantiate_contract_get_app(vec![eth_addr_str], 10000, 1, minter_address);
+
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+    instantiate_contract(
+        vec![eth_addr_str],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
 
     let stargaze_wallet_02 = Addr::unchecked(STARGAZE_WALLET_02);
     let balances = app
@@ -333,9 +377,17 @@ fn test_one_eth_claim_two_stargaze_addresses_invalid() {
         .to_string();
     let eth_address = wallet_1.address();
     let eth_addr_str_1 = format!("{:?}", eth_address);
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app =
-        instantiate_contract_get_app(vec![eth_addr_str_1.clone()], 10000, 1, minter_address);
+
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+    instantiate_contract(
+        vec![eth_addr_str_1.clone()],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
 
     // claim with eth address 1, stargaze wallet 1
     let claim_message = ExecuteMsg::ClaimAirdrop {
@@ -347,14 +399,14 @@ fn test_one_eth_claim_two_stargaze_addresses_invalid() {
         claim_message,
         &mut app,
         stargaze_wallet_01,
-        airdrop_contract,
+        airdrop_contract.clone(),
     )
     .unwrap();
 
     let expected_attributes = [
         Attribute {
             key: "_contract_addr".to_string(),
-            value: "contract0".to_string(),
+            value: AIRDROP_CONTRACT.to_string(),
         },
         Attribute {
             key: "claimed_amount".to_string(),
@@ -385,7 +437,7 @@ fn test_one_eth_claim_two_stargaze_addresses_invalid() {
         eth_address: eth_addr_str_1,
         eth_sig: eth_sig_str_2,
     };
-    let airdrop_contract = Addr::unchecked(AIRDROP_CONTRACT);
+
     let res_2 = execute_contract_with_msg(
         claim_message,
         &mut app,
@@ -397,7 +449,7 @@ fn test_one_eth_claim_two_stargaze_addresses_invalid() {
     let expected_attributes = [
         Attribute {
             key: "_contract_addr".to_string(),
-            value: "contract0".to_string(),
+            value: AIRDROP_CONTRACT.to_string(),
         },
         Attribute {
             key: "claimed_amount".to_string(),
