+++ contracts/sg-eth-airdrop/src/testing/tests/test_eth_whitelist.rs
@@ -1,6 +1,7 @@
 use crate::msg::{ExecuteMsg, QueryMsg};
-use crate::tests_folder::constants::{AIRDROP_CONTRACT, MINTER_CONTRACT, OWNER};
-use crate::tests_folder::shared::instantiate_contract_get_app;
+use crate::tests_folder::claim_constants::{AIRDROP_CONTRACT, OWNER};
+use crate::tests_folder::setup_contracts::{custom_mock_app, instantiate_contract};
+use crate::tests_folder::setup_minter::configure_minter_with_whitelist;
 use cosmwasm_std::Addr;
 use cw_multi_test::Executor;
 
@@ -11,10 +12,20 @@ fn test_instantiate_with_addresses() {
         "addr2".to_string(),
         "addr3".to_string(),
     ];
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let app = instantiate_contract_get_app(addresses, 10000, 1, minter_address);
-    let sg_eth_addr = Addr::unchecked(AIRDROP_CONTRACT);
 
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+
+    instantiate_contract(
+        addresses,
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
+
+    let sg_eth_addr = Addr::unchecked(AIRDROP_CONTRACT);
     let query_msg = QueryMsg::AirdropEligible {
         eth_address: "addr1".to_string(),
     };
@@ -36,8 +47,17 @@ fn test_instantiate_with_addresses() {
 
 #[test]
 fn test_not_authorized_add_eth() {
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app = instantiate_contract_get_app(vec![], 10000, 1, minter_address);
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+
+    instantiate_contract(
+        vec![],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
 
     let fake_admin = Addr::unchecked("fake_admin");
     let sg_eth_addr = Addr::unchecked(AIRDROP_CONTRACT);
@@ -53,8 +73,18 @@ fn test_not_authorized_add_eth() {
 
 #[test]
 fn test_authorized_add_eth() {
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app = instantiate_contract_get_app(vec![], 10000, 1, minter_address);
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+
+    instantiate_contract(
+        vec![],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
+
     let sg_eth_addr = Addr::unchecked(AIRDROP_CONTRACT);
 
     let eth_address = Addr::unchecked("testing_addr");
@@ -68,8 +98,17 @@ fn test_authorized_add_eth() {
 
 #[test]
 fn test_add_eth_and_verify() {
-    let minter_address = Addr::unchecked(MINTER_CONTRACT);
-    let mut app = instantiate_contract_get_app(vec![], 10000, 1, minter_address);
+    let mut app = custom_mock_app();
+    let (minter_addr, _, _, _, _) = configure_minter_with_whitelist(&mut app);
+
+    instantiate_contract(
+        vec![],
+        10000,
+        5,
+        minter_addr,
+        Addr::unchecked(OWNER),
+        &mut app,
+    );
     let sg_eth_addr = Addr::unchecked(AIRDROP_CONTRACT);
 
     let eth_address_str = Addr::unchecked("testing_addr").to_string();
