+++ contracts/whitelist/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod contract;
 mod error;
+pub mod helpers;
 pub mod msg;
 pub mod state;
-
 pub use crate::error::ContractError;
